Fix missing comma, dedupe PDF doc, unify "Issue" and "Add New" to "Send"
